move attributes around
